Add base condition
`FindLastEntry` is a recursive function and without a base exit
condition it will run forever. Added a `max retry`, which upon exhausted
the function will return empty string.

Also added the base directory to perform search on instead of hard coded
path to make the function generic to be used in other places like `ooo`